roll back fungible
