+++ packages/tokens/fungible/src/extensions/mod.rs
@@ -1,3 +1,4 @@
 pub mod burnable;
 pub mod capped;
 pub mod metadata;
+pub mod mintable;
