+++ packages/tokens/fungible/src/lib.rs
@@ -74,11 +74,11 @@ mod fungible;
 mod impl_token_interface_macro;
 mod storage;
 
-pub use extensions::{burnable, capped, metadata};
+pub use extensions::{burnable, capped, metadata, mintable};
 pub use fungible::{emit_approve, emit_transfer, FungibleToken, FungibleTokenError};
 pub use storage::{
-    allowance, allowance_data, approve, balance, mint, set_allowance, spend_allowance,
-    total_supply, transfer, transfer_from, update, AllowanceData, AllowanceKey, StorageKey,
+    allowance, allowance_data, approve, balance, set_allowance, spend_allowance, total_supply,
+    transfer, transfer_from, update, AllowanceData, AllowanceKey, StorageKey,
 };
 
 mod test;
