+++ packages/tokens/fungible/test_snapshots/test/mint_works.1.json
@@ -110,7 +110,7 @@
             },
             "ext": "v0"
           },
-          4095
+          120960
         ]
       ],
       [
@@ -131,7 +131,7 @@
             },
             "ext": "v0"
           },
-          4095
+          120960
         ]
       ]
     ]
